refactor: parser 기능 InputView에서 사용하도록 변경
